LayoutTests/fast/css/counters/deep-before-expected.txt

PASS layoutTestController.counterValueForElementById('div1') is '0'

TEST COMPLETE
